requirements.txt
accelerate==1.12.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
asttokens==3.0.1
async-timeout==4.0.3
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.19
decorator==5.2.1
Deprecated==1.3.1
distro==1.9.0
effdet==0.4.1
emoji==2.15.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
filelock==3.20.1
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-vision==3.11.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
langchain==1.2.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.2
langchain-ollama==1.0.1
langchain-openai==1.1.4
langchain-text-splitters==1.1.0
langdetect==1.0.9
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.5.0
lxml==6.0.2
Markdown==3.10
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.8
matplotlib-inline==0.2.1
ml_dtypes==0.5.4
mpmath==1.3.0
msoffcrypto-tool==5.4.2
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.14.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.2
numpy==2.2.6
olefile==0.47
ollama==0.6.1
omegaconf==2.3.0
onnx==1.20.0
onnxruntime==1.23.2
openai==2.13.0
opencv-python==4.12.0.88
openpyxl==3.1.5
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
pdf2image==1.17.0
pdfminer.six==20251107
pgvector==0.4.2
pi_heif==1.1.1
pikepdf==10.0.2
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
psutil==7.1.3
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.11
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pypandoc==1.16.2
pyparsing==3.2.5
pypdf==6.4.2
pypdfium2==5.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-iso639==2025.11.16
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.45
stack-data==0.6.3
streamlit==1.52.1
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
timm==1.0.22
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.18.21
unstructured-client==0.42.5
unstructured.pytesseract==0.3.15
unstructured_inference==1.1.2
urllib3==2.6.2
uuid_utils==0.12.0
watchdog==6.0.0
wcwidth==0.2.14
webencodings==0.5.1
wrapt==2.0.1
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
